p/excluder: Force line_scaler to return integers
Fixes #1306

+++ printrun/excluder.py
@@ -76,11 +76,11 @@ class ExcluderWindow(gviz.GvizWindow):
 
     def _line_scaler(self, orig):
         x0, y0 = self.gcode_to_real(orig[0], orig[1])
-        x0 = self.p.scale[0] * x0 + self.p.translate[0]
-        y0 = self.p.scale[1] * y0 + self.p.translate[1]
+        x0 = int(self.p.scale[0] * x0 + self.p.translate[0])
+        y0 = int(self.p.scale[1] * y0 + self.p.translate[1])
         x1, y1 = self.gcode_to_real(orig[2], orig[3])
-        x1 = self.p.scale[0] * x1 + self.p.translate[0]
-        y1 = self.p.scale[1] * y1 + self.p.translate[1]
+        x1 = int(self.p.scale[0] * x1 + self.p.translate[0])
+        y1 = int(self.p.scale[1] * y1 + self.p.translate[1])
         width = max(x0, x1) - min(x0, x1) + 1
         height = max(y0, y1) - min(y0, y1) + 1
         return (min(x0, x1), min(y0, y1), width, height,)
